requirements.txt
django-pages-cms>=0.1.20
django-embed-video==1.0.0
